move web to sidecar/api
