+++ python/main.py
@@ -46,10 +46,9 @@ if __name__ == "__main__":
         param_obj = args_type.parse_obj(param_json)
 
     if args.command == "serve":
-        from web import serve
+        from sidecar.api import serve
 
         serve(host="0.0.0.0", port=1487)
 
     else:
-        raise NotImplementedError(f"Command {args.command} is not implemented.")
         raise NotImplementedError(f"Command {args.command} is not implemented.")
